# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.MemberRole:
    enum:
    - owner
    - admin
    - manager
    - member
    - viewer
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleManager
    - RoleMember
    - RoleViewer
  domain.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/domain.TenantStatus'
      updated_at:
        type: string
    type: object
  domain.TenantStatus:
    enum:
    - active
    - suspended
    type: string
    x-enum-varnames:
    - TenantStatusActive
    - TenantStatusSuspended
...
  handler.acceptInvitationRequest:
    properties:
      token:
        type: string
    type: object
  handler.createTenantRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
  handler.inviteMemberRequest:
    properties:
      email:
        type: string
      role:
        $ref: '#/definitions/domain.MemberRole'
    type: object
...
  handler.switchTenantRequest:
    properties:
      tenant_id:
        type: string
    type: object
...
  handler.updateTenantMetadataRequest:
    properties:
      metadata:
        additionalProperties: true
        type: object
    type: object
  handler.updateTenantRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saaskit.io
    name: SaaSKit Support
    url: https://github.com/medaminerjb/saas-kit/issues
  description: Enterprise-grade Identity and Access Management platform with OIDC,
    multi-tenant support, and API key management.
  license:
    name: MIT
    url: https://github.com/medaminerjb/saas-kit/blob/main/LICENSE
  termsOfService: https://github.com/medaminerjb/saas-kit/blob/main/TERMS.md
  title: SaaSKit API
  version: "1.0"
paths:
...
  /tenants:
    get:
      consumes:
      - application/json
      description: Get all organizations the user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: List of tenants
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Create a new organization/tenant
      parameters:
      - description: Tenant details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tenant created
          schema:
            $ref: '#/definitions/domain.Tenant'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create tenant
      tags:
      - tenants
  /tenants/{tenantID}:
    get:
      consumes:
      - application/json
      description: Get details of a specific organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tenant details
          schema:
            $ref: '#/definitions/domain.Tenant'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not a member
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tenant not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get tenant
      tags:
      - tenants
    patch:
      consumes:
      - application/json
      description: Update organization details
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: Tenant updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated tenant
          schema:
            $ref: '#/definitions/domain.Tenant'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update tenant
      tags:
      - tenants
...
  /tenants/{tenantID}/members:
    get:
      consumes:
      - application/json
      description: Get all members of an organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of members
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not a member
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List tenant members
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Invite a user to join an organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: Invitation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.inviteMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invitation created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Invite member
      tags:
      - tenants
  /tenants/{tenantID}/members/{userID}:
    delete:
      consumes:
      - application/json
      description: Remove a member from an organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Member removed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Remove member
      tags:
      - tenants
  /tenants/{tenantID}/metadata:
    get:
      consumes:
      - application/json
      description: Get metadata for an organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tenant metadata
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not a member
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tenant not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get tenant metadata
      tags:
      - tenants
    patch:
      consumes:
      - application/json
      description: Update metadata for an organization
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: Metadata updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateTenantMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated metadata
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update tenant metadata
      tags:
      - tenants
  /tenants/invitations/accept:
    post:
      consumes:
      - application/json
      description: Accept an invitation to join an organization
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.acceptInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Accept invitation
      tags:
      - tenants
  /tenants/switch:
    post:
      consumes:
      - application/json
      description: Switch the currently active organization
      parameters:
      - description: Tenant ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.switchTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Switch successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Switch active tenant
      tags:
      - tenants
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"